+++ ys_codebase/modules/knowledge-db/scripts/__init__.py
@@ -1,3 +1,0 @@
-"""
-knowledge-db scripts package
-"""